+++ config/webpack/plugin/resolve/DirectorySwitcherPlugin.js
