+++ config/webpack/plugin/resolve/LocalPackageAliasPlugin.js
