+++ config/webpack/plugin/resolve/README.md
