+++ tests/config/webpack/plugin/resolve/packages/test/package.json
